Change the particle list to be a grid.
I'm not sure if this is better. It does mean that performance will be
more deterministic since you always loop across all particles in the
grid.

I was hoping this would fix the water simulation from moving around so
much because there is now strict left to right, top to bottom, execution
on the particles, as opposed to it being whichever particle was created
first. But alas, the water simulation is still acting weird.

Performance does appear to be better. I had to put a limiter on the
particle creation code otherwise sometimes it creates two at once.